Add common sections to pom.xml for deployment to OSS Sonatype
We will start deploying to OSS Sonatype (which is mirrored by Maven
Central, i.e. available without having to configure extra <repository>
entries in the pom.xml).

This project is not related to pom-scijava other than being used by
components in the larger SciJava ecosystem, therefore it would be
inappropriate to add pom-scijava as a parent. However, we can copy the
relevant parts of pom-scijava (https://github.com/scijava/pom-scijava),
and that is exactly what this commit does.

It also adds the section declaring that we want to use the
nexus-staging-maven-plugin, to offer a cherry-pickable changeset for
other, similar projects we want to deply to OSS Sonatype.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>